pyproject.toml
[tool.poetry]
name = "xrpl-validator-dashboard"
version = "0.1.0"
description = "Command line UI for XRP Ledger validator operators"
authors = ["Mayukha Vadari <mvadari@ripple.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.8 <3.11"
python-dotenv = "^0.19.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
